requirements.txt
Faker==30.8.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
